исправил ошибку при удалении книги

+++ app/controllers/edumaterials_controller.rb
@@ -56,14 +56,14 @@ class EdumaterialsController < ApplicationController
       # делаем для видео
       @video = Video.find(params[:id])
       link = Coursetovideo.where("video_id = ?", @video.id)
-      link.destroy
+      link.destroy if link.size == 1
       @video.destroy
       flash[:success] = "Видео удалено."
       redirect_to videos_path
     elsif mtype == "Book"
       @book = Book.find(params[:id])
       link = Coursetobook.where("book_id = ?", @book.id)
-      link.destroy
+      link.destroy if link.size == 1
       @book.destroy
       flash[:success] = "Книга удалена."
       redirect_to books_path
